requirements.txt
asgiref==3.2.7
astroid==2.3.3
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.3.0
django-crispy-forms==1.9.0
django-tinymce==3.0.2
djangorestframework==3.11.0
Faker==4.1.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Ming==0.9.2
more-itertools==8.2.0
packaging==20.3
Pillow==7.1.2
pluggy==0.13.1
plyer==1.4.3
psycopg2==2.8.5
py==1.8.1
pycodestyle==2.6.0
pymongo==3.7.2
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
wcwidth==0.1.9
whitenoise==5.1.0
wincertstore==0.2
wrapt==1.11.2
